test(e2e): re-source recentEntries E2E from Backend, retiring the tubafrenzy mirror round-trip
Post-Phase-3 of the tubafrenzy turndown (WXYC/wiki#88), Backend-Service serves GET /playlists/recentEntries from its own Postgres flowsheet table (BS#1860) rather than proxying tubafrenzy's SSE. The former mirror.test.ts asserted the opposite, now-retired direction — a Backend write mirrored to tubafrenzy's recentEntries — which is the exact propagation the mirror-off removes. Once the mirror is off that test would only ever silently skip, reading as "still covered" when it is not. It was already effectively always-skipped in the bs-lml-gate run (no E2E_TUBAFRENZY_URL there, so tubafrenzy is unreachable and the case self-skips).

Replace it with recent-entries.test.ts, which verifies the post-flip invariant: a write made through Backend (POST/PATCH /flowsheet) surfaces on Backend's own /playlists/recentEntries within the polling window. Add a credential-free case guarding the v=1 flat array wire format + X-Last-Modified (the Android contract, BS#1866) — previously no E2E-tier test hit Backend's recentEntries at all. This doubles as the automated guardrail for wiki#88's headline acceptance criterion ("mirror off; dj-site writes still land in Postgres and surface within the polling window").

Drop the now-dead E2E_TUBAFRENZY_URL / tubafrenzyUrl config (the mirror test was its only consumer) and fix the DTO import to the current barrel path (../src/dtos/index.js); the old per-DTO flowsheet.dto.js path no longer exists.
